fix(box_db): fixing the logic of processing deleted records
The system of counting and processing deleted entries in the index has been redesigned. Records are now grouped by ID, which allows you to correctly identify deleted records and select the most recent active versions for each ID. Fixed the error of counting deleted records when there are multiple versions of a single record.