refactor(core): replace direct use of SharedPreferences with abstraction AppPreferences
feat(home): adding a button to go to settings in AppBar for mobile and desktop versions
refactor(titlebar): improving logic of closing the database before application termination
feat(store): adding a getter for the current state of the database
chore(main): adding initialization of AppPreferences and logging of startup information
feat(router): adding a route for the settings screen

The changes include refactoring of working with application settings through a single interface AppPreferences instead of direct use of SharedPreferences. Added new functionality - a button to go to settings on the main screen. Improved logic of safely closing the database when the application terminates.